requirements.txt
alembic==1.13.3
annotated-types==0.7.0
appdirs==1.4.4
arrow==1.3.0
blinker==1.8.2
boto3==1.35.44
botocore==1.35.44
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flexcache==0.3
flexparser==0.3.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==3.0.1
packaging==24.1
pillow==11.0.0
Pint==0.24.3
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
SQLAlchemy==2.0.36
stravalib==2.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
